No bugs and stats are assigned properly to the two bubbles.
Unfortunately, when switching screens from arena to creation, we lose the reference to the winning bubble, despite being in global scope. as a result we've lost our initial game design where the player continuously challenges the winner. As a back up, the player will always fight a fresh random opposing bubble, so there is a working game loop. Additionally the way bubble data is passed around is inelegant and messy. I think I could've afforded to learn a little more on how GDScript handles data via reference/value, however the documentation I found on it was limited.

Also bubble stats are simply applied to the existing bubbles in the fight scene. Initially wanted to dynamically instantiate bubble prefabs into the scene so this project could potentially scale up to include a battle royale-esque mode with a multitude of bubble. Unfortunately I also ran into issues with that, so my attempt is simply commented out in combat manager.